server/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
python-multipart
qrcode[pil]
python-jose[cryptography]
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
dotenv